Disabling NERDTree and easy-tags.
I don't really feel the need to use NERDTree as of now.
easy-tags with the recursive operation takes up a lot of CPU, because of the
ctags command.
Going to use at post-commit hook to run ctags to generate the tag file
and load that tag file via some vim config. This will remove the dependency in easy-tags.
Thanks @vedharish.